Revert "Created a new type for entries. (#497)"
This reverts commit 43f5db3e7e6364f6963076e1eaecc9190bbb60a9.
