+++ Card/Operation/index.ts
@@ -16,15 +16,23 @@ export namespace Operation {
 			created: isoly.DateTime.now(),
 		}
 	}
-	export function fromEntry(entry: Entry | Entry.Failed): Operation | undefined {
-		return entry.status == "failed"
+	export function fromEntry(entry: Entry): Operation | undefined {
+		return entry.type == "unknown"
 			? undefined
 			: {
 					type: "authorization",
-					id: entry.transaction ?? "unknown",
-					status: entry.type == "capture" ? "captured" : "refunded",
+					id: (entry.type != "refund" ? entry.authorization?.id : entry.transaction?.id) ?? "unknown",
+					status: Operation.fromEntryStatus(entry.type),
 					created: isoly.DateTime.now(),
 			  }
+	}
+	export function fromEntryStatus(status: Exclude<Entry.Type, "unknown">): OperationAuthorization.Status {
+		const statusConverter: Record<Exclude<Entry.Type, "unknown">, OperationAuthorization.Status> = {
+			capture: "captured",
+			cancel: "cancelled",
+			refund: "refunded",
+		}
+		return statusConverter[status]
 	}
 	export const type = isly.union(Card.type, OperationAuthorization.type)
 }
