+++ Operation/Creatable.ts
@@ -36,7 +36,7 @@ export namespace Creatable {
 		changes: Changes.type,
 		type: isly.string(types),
 	})
-	export function fromRefund(account: string, settlement: string, entry: Settlement.Entry.Creatable.Refund): Creatable {
+	export function fromRefund(account: string, settlement: string, entry: Settlement.Entry.Refund.Creatable): Creatable {
 		// The Entry.Refund.Creatable has negative amount and fee
 		// The operation amounts should always be positive
 		const [currency, entryAmount] = entry.amount
