+++ Settlement/Entry/Cancel.ts
@@ -1,10 +1,10 @@
 import { isoly } from "isoly"
 import { isly } from "isly"
-import { Amount } from "../../../Amount"
-import { Authorization } from "../../../Authorization"
-import { Batch } from "../../Batch"
-import { Fee } from "../../Fee"
-import { Identifier as SettlementIdentifier } from "../../Identifier"
+import { Amount } from "../../Amount"
+import { Authorization } from "../../Authorization"
+import { Identifier as SettlementIdentifier } from "../../Settlement/Identifier"
+import { Batch } from "../Batch"
+import { Fee } from "../Fee"
 
 export interface Cancel extends Omit<Cancel.Creatable, "settlement"> {
 	status: "succeeded" | "failed"
