+++ Settlement/Entry/Refund.ts
@@ -1,13 +1,13 @@
 import { isoly } from "isoly"
 import { isly } from "isly"
-import { Acquirer } from "../../../Acquirer"
-import { Amount } from "../../../Amount"
-import { Authorization } from "../../../Authorization"
-import { Merchant } from "../../../Merchant"
-import { Transaction } from "../../../Transaction"
-import { Batch } from "../../Batch"
-import { Fee } from "../../Fee"
-import { Identifier as SettlementIdentifier } from "../../Identifier"
+import { Acquirer } from "../../Acquirer"
+import { Amount } from "../../Amount"
+import { Authorization } from "../../Authorization"
+import { Merchant } from "../../Merchant"
+import { Identifier as SettlementIdentifier } from "../../Settlement/Identifier"
+import { Transaction } from "../../Transaction"
+import { Batch } from "../Batch"
+import { Fee } from "../Fee"
 
 export interface Refund extends Omit<Refund.Creatable, "settlement"> {
 	status: "succeeded" | "failed"
