--- Settlement/Entry/fromCreatable.ts
@@ -1,38 +1,0 @@
-import { gracely } from "gracely"
-import { isoly } from "isoly"
-import { Card } from "../../Card"
-import { Rail } from "../../Rail"
-import { Transaction } from "../../Transaction"
-import type { Entry } from "."
-import { Creatable } from "./Creatable"
-
-export function fromCreatable(
-	creatable: Creatable,
-	transaction: Transaction.CardTransaction | gracely.Error | string,
-	card?: Card
-): Entry {
-	let result: Entry
-	const reasons: string[] = []
-	const created = isoly.DateTime.now()
-	if (creatable.type == "unknown")
-		reasons.push("Unknown entry type.")
-	if (gracely.Error.is(transaction))
-		reasons.push(`gracely error: ${JSON.stringify(transaction)}`)
-	else if (typeof transaction == "string")
-		reasons.push(transaction || "No reason provided.")
-	else if (transaction.status != "finalized")
-		reasons.push(`Transaction ${transaction.id} on account ${transaction.accountId} unable to be finalized.`)
-	if (reasons.length > 0)
-		result = { status: "failed", reason: reasons.join("\n"), ...creatable, created }
-	else if (!card)
-		result = { status: "failed", reason: "Missing card", ...creatable, created }
-	else
-		result = {
-			status: "succeeded",
-			...(creatable as Creatable.Known),
-			created,
-			card: Rail.Address.Card.from(card),
-			transaction: (transaction as Transaction.CardTransaction).id,
-		}
-	return result
-}
