+++ Transaction/Incoming.ts
@@ -26,7 +26,7 @@ export namespace Incoming {
 		reference: TransactionReference.type.optional(),
 	})
 
-	export function fromRefund(entry: Settlement.Entry.Creatable.Refund, card: Rail.Address.Card): Incoming {
+	export function fromRefund(entry: Settlement.Entry.Refund.Creatable, card: Rail.Address.Card): Incoming {
 		const [currency, amount] = entry.amount
 		return {
 			account: card,
