+++ Warning/Settlement/UnknownEntry.ts
@@ -2,16 +2,14 @@ import { cryptly } from "cryptly"
 import { isoly } from "isoly"
 import { isly } from "isly"
 import { Authorization } from "../../Authorization"
-import { Entry } from "../../Settlement/Entry"
+import { Unknown } from "../../Settlement/Entry/Unknown"
 import { Identifier } from "../../Settlement/Identifier"
-import { Transaction } from "../../Transaction"
 import { Base } from "../Base"
 
 export interface UnknownEntry extends Base {
 	type: "unknown-entry"
 	resource: Identifier
 	authorization?: Authorization["id"]
-	transaction?: Transaction["id"]
 }
 
 export namespace UnknownEntry {
@@ -19,13 +17,12 @@ export namespace UnknownEntry {
 		type: isly.string("unknown-entry"),
 		resource: Identifier.type,
 		authorization: isly.fromIs("Authorization.id", cryptly.Identifier.is).optional(),
-		transaction: isly.string().optional(),
 	})
-	export function create(entry: Extract<Entry.Failed, { type: "unknown" }>, resource: Identifier): UnknownEntry {
+	export function create(entry: Unknown, resource: Identifier): UnknownEntry {
 		return {
 			type: "unknown-entry",
 			resource,
-			transaction: entry.transaction,
+			authorization: entry.authorization?.id,
 			date: isoly.Date.now(),
 		}
 	}
